Temp fix for "Warning: findDOMNode is deprecated in StrictMode."
Use unsafe_CreateMuiStrictModeTheme instead of createMuiTheme in order to avoid a warning.
The warning is triggered from react-transition-group, and has been fixed in v4.4.0 however, MaterialUI v4.11.1 seems not to be using it yet.
Therefore, there are two ways of getting rid of it: either using an unsafe mui, or Material v5.0 which is not stable yet.
I opted for unsafe mui fix, and only using it in dev, not in prod, for the StrictMode warning is only visible in dev, not in prod.
This fix will hopefully be removed when we upgrade to MaterialUI v5.0

See https://github.com/mui-org/material-ui/issues/13394 for more info on
Warning: findDOMNode is deprecated in StrictMode.