fix: fix console warning regarding undefined required prop
During transition between scenarios with different run templates, the default
value for a given scenario parameter may temporarily not be available. This leads
to console warnings because the defaultParameterValue prop was decalred as required.
Yet, it does not need to be required: the component could use null or undefined as a
default value, or wait for the actual default value to be retrieved during
transitions between scenarios.