fix: [PROD-12176] add new lines at end of table when rows sorting is enabled
Known issue: when the table is sorted and new lines are added, these
lines may appear somewhere in the middle of the other rows when the
sort is disabled. This does not impact the removal of lines (no risk
of accidental data loss), and when exporting, the new lines are
located at the end of the table, as expected. Changing to another
parameters tab or refreshing the page will fix the glitch.