fix: fix dataset disappearing on workspace switch
When creating a new dataset, the link to the new dataset was correctly
saved in the 'current workspace' object in redux, but not in the same
workspace in the 'workspace list' in redux. This lead to new linked
datasets being hidden when visiting the workspace selector and switching
back to the same workspace, requiring a page refresh to show them again.