fix: change behavior when no datasets match the workspace filter
The previous behavior was to show all 'main' datasets in organization
when the workspace filter had no matching results. Yet, this case can
happen frequently, for instance when no datasets have been shared with
the webapp user: the new behavior is to show an empty dataset list in
this case.