fix: fix runner logs parsing to keep only logs of main container
This commit fixes an error during runner logs download when the number
of streams in logs was lower than 2 (e.g. when stderr was empty), and
fixes undesired logs (i.e. not from main container) in the downloaded
file when the order of streams in the API response was not the order
expected by the webapp.