build: add CAFE option to skip CSP generation at build time
When env variable BUILD_TYPE is set to 'cafe', CSP won't be generated in
index.html. This can be necessary when building a generic webapp server
image, that will be later configured by a configuration patch script
(this script will be reponsible to generate and inject CSP before
starting the server).