fix: remove dataset & scenario sharing button for viewer role
Since versions 3.3.2 and 4.1.0, the Cosmo Tech API has disabled the possibility
for the role "viewer" to enumerate users and their roles when fetching resources
and their security data.

This commit hides the dataset & scenario sharing buttons for users when their
role on the resource is "viewer", to prevent having a dialog with an empty
access-control list.